@@ -23,13 +23,13 @@
   ],
   "private": true,
   "dependencies": {
-    "@angular/common": "^13.4.0",
-    "@angular/compiler": "^13.4.0",
-    "@angular/core": "^13.4.0",
-    "@angular/forms": "^13.4.0",
-    "@angular/platform-browser": "^13.4.0",
-    "@angular/platform-browser-dynamic": "^13.4.0",
-    "@angular/router": "^13.4.0",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "@ethereumjs/common": "^2.5.0",
     "@ethereumjs/tx": "^3.3.2",
     "@ethereumjs/vm": "^5.5.3",
@@ -73,10 +73,9 @@
     "zone.js": "~0.11.8"
   },
   "devDependencies": {
-    "@angular/cli": "^13.3.11",
-    "@angular/compiler-cli": "~13.4.0",
-    "@angular/core": "^13.4.0",
-    "@angular/language-service": "~13.4.0",
+    "@angular/cli": "^14.2.12",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/language-service": "^14.3.0",
     "@types/amplitude-js": "^7.0.1",
     "@types/jest": "^26.0.23",
     "@types/lodash": "^4.14.168",
