@@ -3,7 +3,7 @@ import {
   BackendApiService,
   ProfileEntryResponse,
 } from 'src/lib/services/backend-api';
-import { FormControl, Validators } from '@angular/forms';
+import { UntypedFormControl, Validators } from '@angular/forms';
 
 export class CreatorCoinTrade {
   static BUY_VERB = 'Buy';
@@ -40,8 +40,8 @@ export class CreatorCoinTrade {
   networkFeeNanos: number = 0;
 
   // Amount shown in the trade creator form component.
-  amount: FormControl;
-  transferRecipient: FormControl = new FormControl(null, [Validators.required]);
+  amount: UntypedFormControl;
+  transferRecipient: UntypedFormControl = new UntypedFormControl(null, [Validators.required]);
   showUsernameError: boolean = false;
   showPubKeyError: boolean = false;
   showCannotSendToSelfError: boolean = false;
